Upgrade dependencies
This includes a bump to controller-runtime v0.15.0 which necessitates a
couple of small changes in how watching resources is configured.

Also v0.15.0 introduces a requirement to declare the subresources when
registering a scheme for the fake client.

Finally, it looks like the client creation changed how it validates the
config data, which leads to a change to make the test break in a
slightly different way when detecting connectivity issues.