Allow removal when referenced secret exists.
In order to prevent Secrets from being left in the cluster after the
GitopsCluster has been removed, we added a finalizer that checked for
the existence of the secret, and prevented it.

This allows the skipping of this functionality if the GitopsCluster has
an annotation "clusters.gitops.weave.works/no-secret-finalizer".